pyproject.toml
[project]
name = "chatterbox"
version = "0.2.0"
description = "Playwright-driven test harness for the Brya chatbot (Sage)"
requires-python = ">=3.11"
dependencies = [
    "playwright==1.58.0",
    "openai>=1.40",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["harness*"]
